Move electrical sizing test to test_layout file
